requirements.txt
Django==4.2.11
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
djangorestframework==3.15.1
gunicorn==22.0.0
mysqlclient==2.2.4
pip==24.0
setuptools==69.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
wheel==0.42.0